+++ backend/controllers/CommentController.php
@@ -6,6 +6,7 @@ use Yii;
 use backend\models\Comment;
 use backend\models\CommentSearch;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -33,7 +34,12 @@ class CommentController extends Controller
      * Change the status to ACTIVE
      * @return mixed
      */
-    public function actionApprove($id){
+    public function actionApprove($id)
+    {
+        if ( !\Yii::$app->user->can('comment-change-status')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         $comment = Comment::findOne($id);
         $comment->status = Comment::STATUS_ACTIVE;
 
@@ -61,6 +67,10 @@ class CommentController extends Controller
      */
     public function actionIndex()
     {
+        if ( !\Yii::$app->user->can('comment-list')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         $searchModel = new CommentSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -90,6 +100,10 @@ class CommentController extends Controller
      */
     public function actionCreate()
     {
+        if ( !\Yii::$app->user->can('comment-create')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         $model = new Comment();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -110,6 +124,10 @@ class CommentController extends Controller
      */
     public function actionUpdate($id)
     {
+        if ( !\Yii::$app->user->can('comment-update')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -130,6 +148,10 @@ class CommentController extends Controller
      */
     public function actionDelete($id)
     {
+        if ( !\Yii::$app->user->can('comment-delete')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         try{
           $model = $this->findModel($id);
           if($model->delete()) {
