+++ backend/controllers/UserController.php
@@ -7,6 +7,7 @@ use common\models\User;
 use backend\models\UserSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
 use yii\web\UploadedFile;
 use common\models\Helper;
@@ -37,6 +38,10 @@ class UserController extends Controller
      */
     public function actionIndex()
     {
+        if ( !\Yii::$app->user->can('user-admin')) {
+            throw new ForbiddenHttpException("Access denied");
+        }
+
         $searchModel = new UserSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -54,8 +59,14 @@ class UserController extends Controller
      */
     public function actionView($id)
     {
+        $model = $this->findModel($id);
+
+        if ( !\Yii::$app->user->can('user-admin') and !\Yii::$app->user->can('user-view', ['user' => $model])) {
+            throw new ForbiddenHttpException("Access denied");
+        }
+
         return $this->render('view', [
-            'model' => $this->findModel($id),
+            'model' => $model,
         ]);
     }
 
@@ -66,6 +77,11 @@ class UserController extends Controller
      */
      public function actionCreate()
      {
+
+         if ( !\Yii::$app->user->can('user-admin')) {
+             throw new ForbiddenHttpException("Access denied");
+         }
+
          $model = new User(['scenario' => 'create']);
 
          if ($model->load(Yii::$app->request->post())) {
@@ -113,6 +129,10 @@ class UserController extends Controller
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
+
+        if ( !\Yii::$app->user->can('user-admin') and !\Yii::$app->user->can('user-update', ['user' => $model])) {
+            throw new ForbiddenHttpException("Access denied");
+        }
 
         if ($model->load(Yii::$app->request->post())) {
             $model->file = UploadedFile::getInstance($model, 'file');
@@ -170,6 +190,10 @@ class UserController extends Controller
      */
     public function actionChangeStatus($id)
     {
+      if ( !\Yii::$app->user->can('user-change-status')) {
+          throw new ForbiddenHttpException("Access denied");
+      }
+
       $model = $this->findModel($id);
 
       if ($model->status === $model::STATUS_DELETED) {
@@ -194,10 +218,25 @@ class UserController extends Controller
      */
     public function actionDelete($id)
     {
+        $model = $this->findModel($id);
+
+        if ( !\Yii::$app->user->can('user-admin')) {
+            throw new ForbiddenHttpException("Access denied");
+        }
+
         try {
-          $model = $this->findModel($id);
+
           if ( $model->delete() ) {
             unlink('img/users/' . $model->photo);
+          } else {
+            $errors = '';
+            foreach ($model->getErrors() as $key => $value) {
+                foreach ($value as $row => $field) {
+                    //Yii::$app->session->setFlash("danger", $field);
+                    $errors .= $field . "<br>";
+                }
+            }
+            Yii::$app->session->setFlash("danger", $errors);
           }
           Yii::$app->session->setFlash('success', Yii::t('app', 'User deleted successfully'));
         } catch(\Exception $e) {
